MAKO-2423 - Create a testpdflibcmd.exe equivalent for IDistiller()
Initial commit for a testpdflibcmd equivalent (makodistillercmd) that uses Mako's IDistiller interface.

The application takes a single argument file that contains a list of commands to run, one per line.  Each command is read in and converted to an appropriate string that is set with IDistiller::setParameter().  All commands mirror those used by testpdflibcmd, so argument files used with testpdflibcmd can be used with makodistillercmd.  Note however that IDistiller currently only supports a small subset of the features available with PDF Library so where commands are not supported the test application will ignore it.  Supported commands are listed by running the application with no arguments, and when running an argument file the application will print to standard out only the commands which it supports.  The default settings when running makodistillercmd should match testpdflibcmd.

In order to simplify things here I've used the std::filesystem API that comes with C++17, but this is mostly used for the addFonts() API where testpdflibcmd uses directories or wildcards to add multiple fonts.  It is possible to disable the std::filesystem API for use with earlier versions of C++ by setting the #define WANT_STD_FILESYSTEM to 0, though when doing so any paths used in the argument file must be absolute and for adding fonts each font must be added with it's own command, for example:

-faD:\myfont1.ttf
-faD:\myfont2.ttf

Instead of:

-fa..\*.ttf

The VS solutions/projects here are configured to use a nuget package as with the other applications in this repository, but right now I don't have VS2019 so it is configured to use the vc14 package.

I've tested the application by building it with the nuget package and running quick tests with each command and comparing the results with PDF Library.  I've also confirmed that the arg files/resources supplied under MAKO-1660 work with makodistillercmd.